refactor(gui-client): mock Tunnel in-process over IPC channel
Move the debug-only mock out of the Tunnel service process and into the
GUI itself. `ipc::connect` for the Tunnel socket now hands the real
Controller an in-memory `tokio::io::duplex` channel served by an
in-process task, so the controller / IPC / UI can be exercised offline
with no root and no second process.

The IPC client transport becomes a boxed trait object (`Box<dyn
IpcStream>`) so the production socket and the mock channel share one type
without an enum or hand-written `AsyncRead`/`AsyncWrite` delegation. The
GUI gains a `--mock` flag (kept separate from `--skip-portal-auth`); the
Tunnel service's `run-debug --mock` is removed.

Related: #13382
