+++ rust/mise.toml
@@ -55,11 +55,11 @@ env = { CARGO_PROFILE_TEST_OPT_LEVEL = "1", PROPTEST_CASES = "10000", PROPTEST_M
 raw = true
 run = "cargo test tunnel_test --features proptest -- --nocapture"
 
-[tasks.tauri-skip-portal-auth]
-description = "Run the GUI client decoupled from the portal (--skip-portal-auth) for rapid UI iteration. Requires the mock Tunnel service running alongside as root: `firezone-client-tunnel run-debug --mock` (e.g. FIREZONE_NO_TELEMETRY=true sudo ./target/debug/firezone-client-tunnel run-debug --mock)"
+[tasks.tauri-mock]
+description = "Run the GUI client for rapid UI iteration with an in-process mock Tunnel service (--mock) and the portal decoupled (--skip-portal-auth). Self-contained: no root, no separate process."
 dir = "gui-client"
 raw = true
-run = "cargo tauri dev -- -- --skip-portal-auth"
+run = "cargo tauri dev -- -- --skip-portal-auth --mock"
 
 # =============================================================================
 # Composite tasks
